Also remove it from hasAllowedPatterns
